+++ spg/color_cluster_cv.py
@@ -1,0 +1,560 @@
+'''
+Name: trait_extract_parallel.py
+
+Version: 1.0
+
+Summary: Extract plant traits (leaf area, width, height, ) by paralell processing 
+    
+Author: suxing liu
+
+Author-email: suxingliu@gmail.com
+
+Created: 2019-09-29
+
+USAGE:
+
+python3 color_cluster_cv.py -p ~/plant-image-analysis/test/ -ft jpg
+
+
+'''
+
+# import the necessary packages
+from os.path import join
+
+import numpy as np
+import cv2
+import os
+import argparse
+import glob
+import utils
+import matplotlib.pyplot as plt
+
+from sklearn.cluster import KMeans
+from sklearn.cluster import MiniBatchKMeans
+from collections import Counter
+from skimage.color import rgb2lab, deltaE_cie76
+
+import seaborn as sns
+import imutils
+from scipy.interpolate import interp1d
+
+from mpl_toolkits.mplot3d import Axes3D
+
+'''
+from colormath.color_objects import sRGBColor, LabColor
+from colormath.color_conversions import convert_color
+from colormath.color_diff import delta_e_cie2000
+'''
+MBFACTOR = float(1<<20)
+
+# generate foloder to store the output results
+def mkdir(path):
+    # import module
+    import os
+ 
+    # remove space at the beginning
+    path=path.strip()
+    # remove slash at the end
+    path=path.rstrip("\\")
+ 
+    # path exist?   # True  # False
+    isExists=os.path.exists(path)
+ 
+    # process
+    if not isExists:
+        # construct the path and folder
+        #print path + ' folder constructed!'
+        # make dir
+        os.makedirs(path)
+        return True
+    else:
+        # if exists, return 
+        #print path+' path exists!'
+        return False
+        
+
+
+def color_cluster_seg(image, args_colorspace, args_channels, args_num_clusters):
+    
+    # Change image color space, if necessary.
+    colorSpace = args_colorspace.lower()
+
+    if colorSpace == 'hsv':
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
+        
+    elif colorSpace == 'ycrcb' or colorSpace == 'ycc':
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2YCrCb)
+        
+    elif colorSpace == 'lab':
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2LAB)
+        
+    else:
+        colorSpace = 'bgr'  # set for file naming purposes
+
+    # Keep only the selected channels for K-means clustering.
+    if args_channels != 'all':
+        channels = cv2.split(image)
+        channelIndices = []
+        for char in args_channels:
+            channelIndices.append(int(char))
+        image = image[:,:,channelIndices]
+        if len(image.shape) == 2:
+            image.reshape(image.shape[0], image.shape[1], 1)
+            
+    (width, height, n_channel) = image.shape
+    
+    #print("image shape: \n")
+    #print(width, height, n_channel)
+    
+ 
+    # Flatten the 2D image array into an MxN feature vector, where M is the number of pixels and N is the dimension (number of channels).
+    reshaped = image.reshape(image.shape[0] * image.shape[1], image.shape[2])
+    
+
+    # Perform K-means clustering.
+    if args_num_clusters < 2:
+        print('Warning: num-clusters < 2 invalid. Using num-clusters = 2')
+    
+    #define number of cluster
+    numClusters = max(2, args_num_clusters)
+    
+    # clustering method
+    kmeans = KMeans(n_clusters = numClusters, n_init = 40, max_iter = 500).fit(reshaped)
+    
+    # get lables 
+    pred_label = kmeans.labels_
+    
+    # Reshape result back into a 2D array, where each element represents the corresponding pixel's cluster index (0 to K - 1).
+    clustering = np.reshape(np.array(pred_label, dtype=np.uint8), (image.shape[0], image.shape[1]))
+
+    # Sort the cluster labels in order of the frequency with which they occur.
+    sortedLabels = sorted([n for n in range(numClusters)],key = lambda x: -np.sum(clustering == x))
+
+    # Initialize K-means grayscale image; set pixel colors based on clustering.
+    kmeansImage = np.zeros(image.shape[:2], dtype=np.uint8)
+    for i, label in enumerate(sortedLabels):
+        kmeansImage[clustering == label] = int(255 / (numClusters - 1)) * i
+    
+    ret, thresh = cv2.threshold(kmeansImage,0,255,cv2.THRESH_BINARY | cv2.THRESH_OTSU)
+    
+    #return thresh
+    
+    nb_components, output, stats, centroids = cv2.connectedComponentsWithStats(thresh, connectivity=8)
+    
+    sizes = stats[1:, -1]
+    
+    nb_components = nb_components - 1
+    
+    min_size = 150 
+    
+    img_thresh = np.zeros([width, height], dtype=np.uint8)
+    
+    #for every component in the image, you keep it only if it's above min_size
+    for i in range(0, nb_components):
+        if sizes[i] >= min_size:
+            img_thresh[output == i + 1] = 255
+    
+    #from skimage import img_as_ubyte
+    
+    #img_thresh = img_as_ubyte(img_thresh)
+    
+    #print("img_thresh.dtype")
+    #print(img_thresh.dtype)
+    
+    return img_thresh
+
+
+def RGB2HEX(color):
+    return "#{:02x}{:02x}{:02x}".format(int(color[0]), int(color[1]), int(color[2]))
+
+
+
+
+def color_quantization(image, mask, save_path, num_clusters):
+    
+    #grab image width and height
+    (h, w) = image.shape[:2]
+    
+    #change the color storage order
+    #image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
+    
+    #image = cv2.cvtColor(image, cv2.COLOR_BGR2LAB)
+
+    #apply the mask to get the segmentation of plant
+    masked_image_BGR = cv2.bitwise_and(image, image, mask = mask)
+    
+    #define result path for labeled images
+    result_img_path = save_path + 'masked.png'
+    cv2.imwrite(result_img_path, masked_image_BGR)
+    
+    # convert the image from the RGB color space to the L*a*b*
+    # color space -- since we will be clustering using k-means
+    # which is based on the euclidean distance, we'll use the
+    # L*a*b* color space where the euclidean distance implies
+    # perceptual meaning
+    #masked_image = cv2.cvtColor(masked_image_BGR, cv2.COLOR_BGR2LAB)
+    masked_image = cv2.cvtColor(masked_image_BGR, cv2.COLOR_BGR2RGB)
+    
+    #reshape the image to be a list of pixels
+    pixels = masked_image.reshape((masked_image.shape[0] * masked_image.shape[1], 3))
+    
+    ############################################################
+    #Clustering process
+    ###############################################################
+    # cluster the pixel intensities
+    clt = MiniBatchKMeans(n_clusters = num_clusters)
+    #clt = KMeans(n_clusters = args["clusters"])
+    clt.fit(pixels)
+
+    #assign labels to each cluster 
+    labels = clt.fit_predict(pixels)
+
+    #obtain the quantized clusters using each label
+    quant = clt.cluster_centers_.astype("uint8")[labels]
+    
+    #reshape the feature vectors to images
+    quant = quant.reshape((h, w, 3))
+    image_rec = pixels.reshape((h, w, 3))
+    
+    #convert from L*a*b* to RGB
+    quant = cv2.cvtColor(quant, cv2.COLOR_RGB2BGR)
+    #quant = cv2.cvtColor(quant, cv2.COLOR_LAB2BGR)
+    
+    #define result path for labeled images
+    result_img_path = save_path + 'cluster_out.png'
+    
+    # save color_quantization results
+    cv2.imwrite(result_img_path, quant)
+    
+
+    counts = Counter(labels)
+    # sort to ensure correct color percentage
+    counts = dict(sorted(counts.items()))
+    
+    center_colors = clt.cluster_centers_
+
+    # We get ordered colors by iterating through the keys
+    ordered_colors = [center_colors[i] for i in counts.keys()]
+    hex_colors = [RGB2HEX(ordered_colors[i]) for i in counts.keys()]
+    rgb_colors = [ordered_colors[i] for i in counts.keys()]
+
+    #print(hex_colors)
+    
+    index_bkg = [index for index in range(len(hex_colors)) if hex_colors[index] == '#000000']
+    
+    #print(index_bkg[0])
+
+    #print(counts)
+    #remove background color 
+    del hex_colors[index_bkg[0]]
+    del rgb_colors[index_bkg[0]]
+    
+    # Using dictionary comprehension to find list 
+    # keys having value . 
+    delete = [key for key in counts if key == index_bkg[0]] 
+  
+    # delete the key 
+    for key in delete: del counts[key] 
+   
+    fig = plt.figure(figsize = (6, 6))
+    plt.pie(counts.values(), labels = hex_colors, colors = hex_colors)
+
+    #define result path for labeled images
+    result_img_path = save_path + 'pie_color.png'
+    plt.savefig(result_img_path)
+        
+    #build a histogram of clusters and then create a figure representing the number of pixels labeled to each color
+    hist = utils.centroid_histogram(clt)
+
+    #remove the background color cluster
+    clt.cluster_centers_ = np.delete(clt.cluster_centers_, index_bkg[0], axis=0)
+    
+    #build a histogram of clusters using center lables
+    numLabels = utils.plot_centroid_histogram(save_path,clt)
+
+    #create a figure representing the distribution of each color
+    bar = utils.plot_colors(hist, clt.cluster_centers_)
+
+    #save a figure of color bar 
+    utils.plot_color_bar(save_path, bar)
+    
+    
+    return rgb_colors
+    
+    
+
+def color_region(image, mask, output_directory, file_name, num_clusters):
+    
+    # read the image
+     #grab image width and height
+    (h, w) = image.shape[:2]
+
+    #apply the mask to get the segmentation of plant
+    masked_image_ori = cv2.bitwise_and(image, image, mask = mask)
+    
+    #define result path for labeled images
+    result_img_path = join(output_directory, f"{file_name}.masked.png")
+    cv2.imwrite(result_img_path, masked_image_ori)
+    
+    # convert to RGB
+    image_RGB = cv2.cvtColor(masked_image_ori, cv2.COLOR_BGR2RGB)
+
+    # reshape the image to a 2D array of pixels and 3 color values (RGB)
+    pixel_values = image_RGB.reshape((-1, 3))
+    
+    # convert to float
+    pixel_values = np.float32(pixel_values)
+
+    # define stopping criteria
+    criteria = (cv2.TERM_CRITERIA_EPS + cv2.TERM_CRITERIA_MAX_ITER, 100, 0.2)
+
+    # number of clusters (K)
+    #num_clusters = 5
+    compactness, labels, (centers) = cv2.kmeans(pixel_values, num_clusters, None, criteria, 10, cv2.KMEANS_RANDOM_CENTERS)
+
+    # convert back to 8 bit values
+    centers = np.uint8(centers)
+
+    # flatten the labels array
+    labels_flat = labels.flatten()
+
+    # convert all pixels to the color of the centroids
+    segmented_image = centers[labels_flat]
+
+    # reshape back to the original image dimension
+    segmented_image = segmented_image.reshape(image_RGB.shape)
+
+
+    segmented_image_BRG = cv2.cvtColor(segmented_image, cv2.COLOR_RGB2BGR)
+    #define result path for labeled images
+    result_img_path = join(output_directory, f"{file_name}.clustered.png")
+    cv2.imwrite(result_img_path, segmented_image_BRG)
+
+
+    '''
+    fig = plt.figure()
+    ax = Axes3D(fig)        
+    for label, pix in zip(labels, segmented_image):
+        ax.scatter(pix[0], pix[1], pix[2], color = (centers))
+            
+    result_file = (save_path + base_name + 'color_cluster_distributation.png')
+    plt.savefig(result_file)
+    '''
+    #Show only one chosen cluster 
+    #masked_image = np.copy(image)
+    masked_image = np.zeros_like(image_RGB)
+
+    # convert to the shape of a vector of pixel values
+    masked_image = masked_image.reshape((-1, 3))
+    # color (i.e cluster) to render
+    #cluster = 2
+
+    
+    clrs = sns.color_palette('husl', n_colors = num_clusters)  # a list of RGB tuples
+
+    color_conversion = interp1d([0,1],[0,255])
+
+
+    for cluster in range(num_clusters):
+
+        print("Processing Cluster{0} ...\n".format(cluster))
+        #print(clrs[cluster])
+        #print(color_conversion(clrs[cluster]))
+
+        masked_image[labels_flat == cluster] = centers[cluster]
+
+        #print(centers[cluster])
+
+        #convert back to original shape
+        masked_image_rp = masked_image.reshape(image_RGB.shape)
+
+        #masked_image_BRG = cv2.cvtColor(masked_image, cv2.COLOR_RGB2BGR)
+        #cv2.imwrite('maksed.png', masked_image_BRG)
+
+        gray = cv2.cvtColor(masked_image_rp, cv2.COLOR_BGR2GRAY)
+
+        # threshold the image, then perform a series of erosions +
+        # dilations to remove any small regions of noise
+        thresh = cv2.threshold(gray, 0, 255, cv2.THRESH_BINARY)[1]
+
+        cnts = cv2.findContours(thresh.copy(), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+        cnts = imutils.grab_contours(cnts)
+        #c = max(cnts, key=cv2.contourArea)
+
+        '''
+        # compute the center of the contour area and draw a circle representing the center
+        M = cv2.moments(c)
+        cX = int(M["m10"] / M["m00"])
+        cY = int(M["m01"] / M["m00"])
+        # draw the countour number on the image
+        result = cv2.putText(masked_image_rp, "#{}".format(cluster + 1), (cX - 20, cY), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (255, 255, 255), 2)
+        '''
+        
+        if not cnts:
+            print("findContours is empty")
+        else:
+            
+            # loop over the (unsorted) contours and draw them
+            for (i, c) in enumerate(cnts):
+
+                #result = cv2.drawContours(masked_image_rp, c, -1, (0, 0, 255), 2)
+                result = cv2.drawContours(masked_image_rp, c, -1, color_conversion(clrs[cluster]), 2)
+
+            #result = result(np.where(result == 0)== 255)
+            result[result == 0] = 255
+
+
+            result_BRG = cv2.cvtColor(result, cv2.COLOR_RGB2BGR)
+            result_img_path = join(output_directory, f"{file_name}.result.{cluster}.png")
+            cv2.imwrite(result_img_path, result_BRG)
+
+    '''
+    result_BRG = cv2.cvtColor(result, cv2.COLOR_RGB2BGR)
+    result_img_path = save_path + 'result_all.png' 
+    cv2.imwrite(result_img_path, result_BRG)
+    '''
+    
+    counts = Counter(labels_flat)
+    # sort to ensure correct color percentage
+    counts = dict(sorted(counts.items()))
+    
+    center_colors = centers
+
+    # We get ordered colors by iterating through the keys
+    ordered_colors = [center_colors[i] for i in counts.keys()]
+    hex_colors = [RGB2HEX(ordered_colors[i]) for i in counts.keys()]
+    rgb_colors = [ordered_colors[i] for i in counts.keys()]
+
+    #print(hex_colors)
+    
+    index_bkg = [index for index in range(len(hex_colors)) if hex_colors[index] == '#000000']
+    
+    #print(index_bkg[0])
+
+    #print(counts)
+    #remove background color 
+    del hex_colors[index_bkg[0]]
+    del rgb_colors[index_bkg[0]]
+    
+    # Using dictionary comprehension to find list 
+    # keys having value . 
+    delete = [key for key in counts if key == index_bkg[0]] 
+  
+    # delete the key 
+    for key in delete: del counts[key] 
+   
+    fig = plt.figure(figsize = (6, 6))
+    plt.pie(counts.values(), labels = hex_colors, colors = hex_colors)
+
+    #define result path for labeled images
+    result_img_path = join(output_directory, f"{file_name}.pie_color.png")
+    plt.savefig(result_img_path)
+
+   
+    return rgb_colors
+    
+    
+
+
+if __name__ == '__main__':
+
+    # construct the argument and parse the arguments
+    ap = argparse.ArgumentParser()
+    ap.add_argument("-p", "--path", required = True, help = "path to image file")
+    ap.add_argument("-ft", "--filetype", required = True, help = "image filetype")
+    args = vars(ap.parse_args())
+    
+    # setting path to model file
+    file_path = args["path"]
+    ext = args['filetype']
+    
+    #accquire image file list
+    filetype = ('*.' + ext)
+    image_file_path = file_path + filetype
+    
+    #accquire image file list
+    imgList = sorted(glob.glob(image_file_path))
+    
+    
+    
+    
+    image_file = imgList[0]
+    
+    abs_path = os.path.abspath(image_file)
+    
+    filename, file_extension = os.path.splitext(abs_path)
+
+    file_size = os.path.getsize(image_file)/MBFACTOR
+    
+    
+     
+    # make the folder to store the results
+    #current_path = abs_path + '/'
+    base_name = os.path.splitext(os.path.basename(filename))[0]
+    print("Exacting traits for image : {0}\n".format(str(base_name)))
+     
+    # save folder construction
+    mkpath = os.path.dirname(abs_path) +'/' + base_name
+    mkdir(mkpath)
+    save_path = mkpath + '/'
+    print ("results_folder: " + save_path)
+    
+    
+    if (file_size > 5.0):
+        print("It will take some time due to larger file size {0} MB".format(str(int(file_size))))
+    else:
+        print("Segmentaing plant object using automatic color clustering method... ")
+    
+
+    image = cv2.imread(image_file)
+    print("Shape: {}".format(image.shape))
+    
+    #make backup image
+    orig = image.copy()
+    
+    ################################
+    r, g, b = cv2.split(orig)
+    r = r.flatten()
+    g = g.flatten()
+    b = b.flatten()
+    
+    fig = plt.figure()
+    ax = Axes3D(fig)
+    ax.scatter(r, g, b)
+    result_file = (save_path + base_name + 'color_distributation.png')
+    plt.savefig(result_file)
+    ################################################
+    
+    args_colorspace = 'lab'
+    args_channels = '1'
+    args_num_clusters = 2
+
+    #color clustering based plant object segmentation
+    thresh = color_cluster_seg(orig, args_colorspace, args_channels, args_num_clusters)
+    
+    # save segmentation result
+    result_file = (save_path + base_name + '_seg' + file_extension)
+    #print(filename)
+    cv2.imwrite(result_file, thresh)
+    
+    
+    num_clusters = 5
+    #save color quantization result
+    #rgb_colors = color_quantization(orig, thresh, save_path, num_clusters)
+    
+    rgb_colors = color_region(orig, thresh, save_path, num_clusters)
+    
+    #print ("List index-value are : ") 
+    
+    
+    selected_color = rgb2lab(np.uint8(np.asarray([[rgb_colors[0]]])))
+    
+    for index, value in enumerate(rgb_colors): 
+        #print(index, value) 
+        
+        curr_color = rgb2lab(np.uint8(np.asarray([[value]])))
+        diff = deltaE_cie76(selected_color, curr_color)
+        print(index, value, diff) 
+        
+    
+
